LayoutTests/inspector/indexeddb/clearObjectStore.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/utilities.js"></script>
<script>
function test()
{
    function sortedObjectStores(database) {
        const stores = database.objectStores;
        stores.sort((a, b) => a.name > b.name);
        return stores;
    }

    let suite = InspectorTest.createAsyncSuite("IndexedDB.clearObjectStore");

    suite.addTestCase({
        name: "ClearDatabases",
        description: "Remove any existing IndexedDB databases.",
        test(resolve, reject) {
            // FIXME: Remove any existing IndexedDB databases that might exist to workaround:
            // <https://webkit.org/b/148006> Each test should run with its own clean data store
            IndexedDBAgent.requestDatabaseNames(WI.networkManager.mainFrame.securityOrigin, (error, names) => {
                InspectorTest.evaluateInPage("deleteDatabaseNames(" + JSON.stringify(names) + ")", resolve);
            });
        }
    });

    suite.addTestCase({
        name: "CreateDatabaseAndClearObjectStore",
        description: "Ensure that clearing an object store clears entries in that object store.",
        test(resolve, reject) {
            InspectorTest.evaluateInPage("createDatabaseWithStores('CompleteDatabase', 1)");
            IndexedDBAgent.requestDatabase(WI.networkManager.mainFrame.securityOrigin, "CompleteDatabase", (error, database) => {
                InspectorTest.expectNoError(error);

                function expectObjectStoreEntryCount(objectStore, expectedCount, message, callback) {
                    const securityOrigin = WI.networkManager.mainFrame.securityOrigin;
                    const indexName = "";
                    const skipCount = 0;
                    const pageSize = 10;
                    IndexedDBAgent.requestData(securityOrigin, database.name, objectStore.name, indexName, skipCount, pageSize, (error, entries, hasMore) => {
                        InspectorTest.expectNoError(error);
                        InspectorTest.expectThat(entries.length === expectedCount, message);
                        callback();
                    });
                }

                let securityOrigin = WI.networkManager.mainFrame.securityOrigin;
                InspectorTest.expectThat(database.objectStores.length === 3, "Database should have 3 object stores.");
                let [emptyObjectStore, reviewersObjectStore, statsObjectStore] = sortedObjectStores(database);
                let instructions = [
                    { type: "check", objectStore: emptyObjectStore, expectedCount: 0, message: "'Empty' object store should always be empty." },
                    { type: "check", objectStore: reviewersObjectStore, expectedCount: 4, message: "'Reviewers' object store should start with 4 entries." },
                    { type: "check", objectStore: statsObjectStore, expectedCount: 4, message: "'Stats' object store should start with 4 entries." },
                    { type: "clear", objectStore: reviewersObjectStore },
                    { type: "check", objectStore: emptyObjectStore, expectedCount: 0, message: "'Empty' object store should always be empty." },
                    { type: "check", objectStore: reviewersObjectStore, expectedCount: 0, message: "'Reviewers' object store should have been cleared." },
                    { type: "check", objectStore: statsObjectStore, expectedCount: 4, message: "'Stats' object store remain unchanged with 4 entries." },
                    { type: "clear", objectStore: statsObjectStore },
                    { type: "check", objectStore: emptyObjectStore, expectedCount: 0, message: "'Empty' object store should always be empty." },
                    { type: "check", objectStore: reviewersObjectStore, expectedCount: 0, message: "'Reviewers' object store should have been cleared." },
                    { type: "check", objectStore: statsObjectStore, expectedCount: 0, message: "'Stats' object store should have been cleared." },
                ];

                function test() {
                    instruction = instructions.shift();
                    if (!instruction) {
                        InspectorBackend.runAfterPendingDispatches(resolve);
                        return;
                    }

                    if (instruction.type == "check")
                        expectObjectStoreEntryCount(instruction.objectStore, instruction.expectedCount, instruction.message, test);
                    else if (instruction.type == "clear") {
                        IndexedDBAgent.clearObjectStore(securityOrigin, database.name, instruction.objectStore.name, (error) => {
                            InspectorTest.expectNoError(error);
                            test();
                        });
                    }
                }

                test();
            });
        }
    });

    suite.addTestCase({
        name: "NoSuchObjectStore",
        test(resolve, reject) {
            InspectorTest.awaitEvent("DatabaseCreated")
            .then((event) => {
                IndexedDBAgent.requestDatabase(WI.networkManager.mainFrame.securityOrigin, "EmptyDatabase", (error, database) => {
                    IndexedDBAgent.clearObjectStore(WI.networkManager.mainFrame.securityOrigin, database.name, "NoSuchObjectStore", (error) => {
                        InspectorTest.expectThat(error, "Should be an error attempting to clear an object store that does not exist.");
                        resolve();
                    });
                });
            });

            InspectorTest.evaluateInPage("createEmptyDatabase('EmptyDatabase', 1)");
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
</body>
</html>
